move Rotor
