+++ src/main/java/titanicsend/pattern/TEPerformancePattern.java
@@ -17,6 +17,7 @@ import titanicsend.pattern.jon.VariableSpeedTimer;
 import titanicsend.pattern.jon._CommonControlGetter;
 import titanicsend.pattern.yoffa.framework.TEShaderView;
 import titanicsend.util.MissingControlsManager;
+import titanicsend.util.Rotor;
 import titanicsend.util.TE;
 import titanicsend.util.TEColor;
 
@@ -752,68 +753,7 @@ public abstract class TEPerformancePattern extends TEAudioPattern {
         }
     }
 
-    /**
-     * Class to support incremental rotation over variable-speed time
-     *
-     * The rate is tied to the engine bpm and the input time value, which is usually
-     * controlled by the variable speed timer associated with the speed or spin controls.
-     * (but anything with a seconds.millis timer can generate rotational angles this way.)
-     */
-    protected class Rotor {
-        private double maxSpinRate = Math.PI;
-        private double angle = 0;
-        private double lastTime = 0;
 
-        // Internal: Called on every frame to calculate and memoize the current
-        // spin angle so that calls to getAngle() during a frame
-        // will always return the same value no matter how long the frame
-        // calculations take.
-        void updateAngle(double time, double ctlValue) {
-            // if this is the first frame, or if the timer was restarted,
-            // we skip calculation for a frame.  Otherwise, do
-            // the incremental angle calculation...
-            if (lastTime != 0) {
-                // calculate change in angle since last frame.
-                // Note: revised calculation restricts maximum speed while still allowing
-                // you to get to maximum speed at slower bpm.
-                double et = Math.min(maxSpinRate, maxSpinRate * (time - lastTime));
-                angle -= et % LX.TWO_PI;
-            }
-            lastTime = time;
-        }
-
-        /**
-         * @return Current rotational angle, either computed, or taken from
-         */
-        double getAngle() {
-            return angle;
-        }
-
-        void setAngle(double angle) {
-            this.angle = angle;
-        }
-
-        void addAngle(double offset) {
-            this.angle += offset;
-        }
-
-        void reset() {
-            angle = 0;
-            lastTime = 0;
-        }
-
-        /**
-         * Sets maximum spin rate for all patterns using this rotor.  Note that a Rotor
-         * object is associated with a timer, which can be a VariableSpeedTimer.  So
-         * "seconds" may be variable in duration, and can be positive or negative.
-         *
-         * @param radiansPerSecond
-         */
-        void setMaxSpinRate(double radiansPerSecond) {
-            maxSpinRate = radiansPerSecond;
-        }
-
-    }
 
     private final VariableSpeedTimer iTime = new VariableSpeedTimer();
     private final VariableSpeedTimer spinTimer = new VariableSpeedTimer();
